+++ LinkedListTest.h
@@ -1,0 +1,6 @@
+class TestTask : public RplTask {
+	void onStart() {};
+	void onResume() {};
+	void onPause() {};
+	void onLoop() {};
+};
